requirements.txt
flask
flask-cors
cryptography
python-dotenv
requests
werkzeug
